requirements.txt
Flask==3.0.0
flask-cors==4.0.1
gunicorn==22.0.0
groq==0.11.0
python-dotenv==1.1.1
python-docx==1.2.0
pdfplumber==0.11.7
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.2.1
itsdangerous==2.2.0
blinker==1.9.0
Pillow==11.3.0
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
numpy==1.26.4
